refactor: slight adjustment to file names and component dir structure
