+++ components/Detail.tsx
