+++ components/EventBtn.tsx
@@ -1,6 +1,7 @@
 import { EventType } from "@lib/types";
+import { Button } from "./ui/Button";
 
-export default function Event({
+export default function EventBtn({
   _id,
   typeIcon,
   date,
@@ -9,9 +10,13 @@ export default function Event({
 }: EventType) {
   return (
     <div key={_id} className="relative flex items-start py-8">
-      <button className="relative text-4xl bg-white dark:bg-gray-700 hover:bg-primary dark:hover:bg-primary border-2 border-primary dark:border-gray-700 dark:hover:border-primary p-4 rounded-lg shadow-md transform hover:scale-[1.15] active:scale-100 [&>svg>path]:hover:text-white cursor-pointer">
+      <Button
+        variant="icon"
+        size="max"
+        className="text-4xl bg-white dark:bg-dm-surface hover:bg-primary dark:hover:bg-primary border-primary p-4 rounded-lg shadow-md hover:scale-[1.15] active:scale-100 [&>svg>path]:hover:text-white"
+      >
         {typeIcon}
-      </button>
+      </Button>
       <div className="px-4">
         <h1 className="text-secondary dark:text-secondary-light text-4xl font-black capitalize">
           {date}
