+++ components/ImageCredits.tsx
