+++ components/Quote.tsx
