+++ components/sections/AboutSection.tsx
@@ -2,10 +2,10 @@ import { useFetch } from "@hooks/useFetch";
 import { AboutType } from "@lib/types";
 import Picture from "@components/common/Picture";
 import Section from "@components/common/Section";
-import Heading from "@components/ui/Heading";
-import AboutUi from "@components/ui/AboutUi";
-import LoadingSpinner from "@components/ui/LoadingSpinner";
-import ImageCredits from "@components/ui/ImageCredits";
+import Heading from "@components/Heading";
+import AboutUi from "@components/AboutUi";
+import LoadingSpinner from "@components/LoadingSpinner";
+import ImageCredits from "@components/ImageCredits";
 
 export default async function AboutSection() {
   const { abouts } = await useFetch("/api/abouts");
@@ -39,7 +39,7 @@ export default async function AboutSection() {
       <div className="about-me w-full flex flex-col lg:flex-row justify-between items-center lg:items-start xl:justify-evenly ">
         <div className="relative w-[300px] h-[300px] lg:w-[400px] lg:h-[400px] 2xl:w-[500px] 2xl:h-[500px] aspect-square shadow-md hover:shadow-primary-dark mb-4 rounded-lg [&>span]:hover:opacity-70 overflow-hidden">
           <Picture
-            src={"about-me-image_aidnjy"}
+            src={`https://res.cloudinary.com/${process.env.NEXT_PUBLIC_CLOUDINARY_CLOUD_NAME}/image/upload/v1710343815/portfolio/about-me-image_aidnjy`}
             alt={"Jam Moonbami"}
             isPriority={true}
           />
